Version 5.4. Added compiler specializations to previously optimized inner loops. "Shuffled" SIMD interpolation code is not efficient on Apple M1. Intel C++ Compiler vectorizes "whole stepping" interpolation as good as a manually-written SSE. Reorganized SIMD instructions for a slightly better performance. Changed internal buffer sizes of half-band resamplers (1-2% performance boost). Fixed compiler warnings in PFFFT code. Added several asserts to the code.

+++ pffft_double/pffft_priv_impl.h
@@ -1818,7 +1818,7 @@ void FUNC_VALIDATE_SIMD_A() {
 
 static void pffft_assert1( float result, float ref, const char * vartxt, const char * functxt, int * numErrs, const char * f, int lineNo )
 {
-  if ( !( fabsf( result - ref ) < 0.01F ) )
+  if ( !( fabs( result - ref ) < 0.01 ) )
   {
     fprintf(stderr, "%s: assert for %s at %s(%d)\n  expected %f  value %f\n", functxt, vartxt, f, lineNo, ref, result);
     ++(*numErrs);
